Put url types in their own area like other types. Updated imports to match new path!
